requirements.txt
Flask
flask-cors
pandas
python-dotenv
requests
openai
parsedatetime
python-dateutil
gunicorn
